pyproject.toml
[project]
name = "extractgraph"
version = "0.1.0"
requires-python = ">=3.11.13"
dependencies = [
    "langextract[openai]>=1.0.9",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "pyvis>=0.3.2",
]
